% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{geocode in Colombia}
\usage{
get_data(
  search,
  survey,
  year,
  acs_variables,
  geography,
  api_key,
  geometry,
  path,
  file_type,
  CRS,
  format,
  moe_level,
  name
)
}
\arguments{
\item{search}{The dataframe containing the information.}

\item{survey}{The CRS for geocoding the addresses.}

\item{year}{The name of the column with the id for keeping track of your data.}

\item{acs_variables}{The name of the column with the addresses for geocoding.}

\item{geography}{The name of the column with the city of the address.}

\item{path}{description}

\item{file_type}{description}

\item{CRS}{description}

\item{format}{description}

\item{moe_level}{description}

\item{name}{description}
}
\value{
The coordinates of the address given, and the match score.
}
\description{
This function allows you to retrieve information from the US Census Bureau
}
\examples{
variables <- c("DP05_0001E", "S1701_C03_001E", "B18101_001E", "B18101_004E", "B18101_007E", "B18101_010E",
               "B18101_013E", "B18101_016E", "B18101_019E", "B18101_023E", "B18101_026E", "B18101_029E",
               "B18101_032E", "B18101_035E", "B18101_038E")

info <- get_data(
  acs_variables = variables
  )
}
\keyword{Colombia}
\keyword{address.}
\keyword{geocode}
