% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Retrieve information for Puerto Rico from the US Census Bureau}
\usage{
get_data(
  search,
  survey,
  year,
  acs_variables,
  geography,
  api_key,
  geometry,
  path,
  file_type,
  CRS,
  format,
  moe_level,
  name,
  state
)
}
\arguments{
\item{search}{(optional) The name of a county to search specific information. If ommited, the information will be for all
Puerto Rico.}

\item{survey}{The ACS survey to be searched. The ACS contains one-year, three-year, and five-year surveys expressed as "acs1", "acs3", and "acs5".}

\item{year}{The year, or endyear, of the ACS sample. As now, 5-year ACS data is available from 2009 through 2021; 1-year ACS data is available from 2005 through 2021, except for 2020.}

\item{acs_variables}{The name of the column with the addresses for geocoding.}

\item{geography}{The name of the column with the city of the address.}

\item{path}{description}

\item{file_type}{description}

\item{CRS}{description}

\item{format}{description}

\item{moe_level}{description}

\item{name}{description}
}
\value{
The coordinates of the address given, and the match score.
}
\description{
This function allows you to retrieve information from the US Census Bureau
}
\examples{
variables <- c("DP05_0001E", "S1701_C03_001E", "B18101_001E", "B18101_004E", "B18101_007E", "B18101_010E",
               "B18101_013E", "B18101_016E", "B18101_019E", "B18101_023E", "B18101_026E", "B18101_029E",
               "B18101_032E", "B18101_035E", "B18101_038E")

info <- get_data(
  acs_variables = variables
  )
}
\keyword{Colombia}
\keyword{address.}
\keyword{geocode}
